requirements/development.txt
asgiref==3.6.0
astroid==2.15.4
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorclass==2.2.2
coverage==7.2.3
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
Django==4.2
django-appconf==1.0.5
django-compressor==4.1
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-kn-defaults==0.3.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djvue==0.2.1
docopt==0.6.2
elastic-apm==6.13.2
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==12.0.1
filelock==3.12.0
flake8==6.0.0
flake8-isort==5.0.3
identify==2.5.23
idna==3.4
imgkit==1.2.3
importlib-resources==5.12.0
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
lazy-object-proxy==1.9.0
limits==3.4.0
mccabe==0.7.0
mixer==7.2.2
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.10
Pillow==9.5.0
pip-upgrader==1.4.15
platformdirs==3.4.0
pluggy==1.0.0
pre-commit==2.20.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pylogbeat==2.0.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-logstash-async==2.5.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.0
requests==2.29.0
rjsmin==1.2.0
sentry-sdk==1.11.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
typing-extensions==4.5.0
urllib3==1.26.15
virtualenv==20.22.0
wrapt==1.15.0
zipp==3.15.0
channels==3.0.5
weasyprint==59.0
redis==4.5.5
